+++ app/controllers/users_controller.rb
@@ -3,6 +3,10 @@ class UsersController < ApplicationController
   before_action :authorize
   skip_before_action :authorize, only: [:create]
 
+  def index
+    render json: User.all, status: :ok
+  end
+  
   def create
     user = User.create(user_params)
     if user.valid?
@@ -36,7 +40,7 @@ class UsersController < ApplicationController
     user = User.find(params[:user_id])
     games = user.games
     if games.size > 0
-      gamePackages = games.map {|game| {game: {**game.attributes, host: game.host.username, players: game.players}, board: game.board} }
+      gamePackages = games.map {|game| game.package}
       render json: gamePackages, status: :ok
     else
       render json: { errors: "Sorry, you have no games" }, status: :not_found
@@ -44,25 +48,33 @@ class UsersController < ApplicationController
   end
 
   def update
-    user = User.find(params[:id])
-    user.update(user_params)
-    if user.valid?
-      render json: user, status: :accepted
+    if params[:id] == session[:user_id]
+      user = User.find(params[:id])
+      user.update(user_params)
+      if user.valid?
+        render json: user, status: :accepted
+      else
+        render json: { errors: user.errors.full_messages }, status: :unprocessable_entity
+      end
     else
-      render json: { errors: user.errors.full_messages }, status: :unprocessable_entity
+      render json: { error: "Not authorized" }, status: :unauthorized
     end
   end
 
   def destroy
-    user = User.find(params[:id])
-    user.destroy
-    head :no_content
+    if params[:id] == session[:user_id]
+      user = User.find(params[:id])
+      user.destroy
+      head :no_content
+    else
+      render json: { error: "Not authorized" }, status: :unauthorized
+    end
   end
 
   private
 
   def user_params
-      params.permit(:username, :password, :email, :pic_url)
+    params.permit(:username, :password, :email, :pic_url)
   end
 
   def authorize
